requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beancount==2.3.5
beautifulsoup4==4.10.0
blip-ci==0.0.3
bottle==0.12.19
cachetools==5.0.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
cheroot==8.6.0
click==8.1.3
clip-interrogator==0.5.4
dateparser==1.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
fairscale==0.4.4
fastapi==0.77.1
fava==1.23.1
filelock==3.9.0
Flask==2.0.3
Flask-Babel==2.0.0
ftfy==6.1.1
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.55.0
h11==0.14.0
httplib2==0.20.4
huggingface-hub==0.12.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.0
jaraco.functools==3.5.0
Jinja2==3.1.2
joblib==1.2.0
khoj-assistant==0.3.0
lxml==4.8.0
markdown-it-py==2.2.0
markdown2==2.4.2
MarkupSafe==2.1.0
mdurl==0.1.2
more-itertools==9.0.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
open-clip-torch==2.19.0
openai==0.20.0
openpyxl==3.1.1
packaging==21.3
pandas==1.5.3
pandas-stubs==1.5.3.230214
pi==0.1.2
Pillow==9.3.0
pluggy==1.0.0
ply==3.11
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
Pygments==2.14.0
pyparsing==3.0.7
PyQt6==6.3.1
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
pytest==7.0.1
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.28.2
rich==13.3.1
rsa==4.8
safetensors==0.3.1
schedule==1.1.0
scikit-learn==1.2.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
starlette==0.19.1
threadpoolctl==3.1.0
timm==0.4.12
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
types-pytz==2022.7.1.2
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.17.6
wcwidth==0.2.6
Werkzeug==2.0.3
